Fixed Stopping and Starting scripts

+++ backend/event/stopEvent.sh
@@ -1,7 +1,9 @@
 # Kill all the processes
 while read line           
-do           
-    kill -9 $line 
+do
+	if ps -p $line > /dev/null; then
+    	kill -9 $line
+	fi
 done <PID
 
 > PID
