+++ backend/restful/startRestful.sh
@@ -1,7 +1,13 @@
+# Recreate Logfile
+> LOG
+touch LOG
+
 # Kill all the processes
 while read line           
-do           
-    kill -9 $line 
+do
+	if ps -p $line > /dev/null; then
+    	kill -9 $line
+	fi
 done <PID
 
 # Clear the PIDs
@@ -12,6 +18,9 @@ services=$(ls . | grep .js)
 
 # Start all the services
 for service in $services; do
-    node $service &
+    node $service &> LOG &
     echo $! >> PID
+    sleep 1
 done
+
+echo "Booted"
